Revert "quartz: Remove X11 ifdefs"
This reverts commit eb81ec945c073bc87dc622d8b27906270953df8b.

We are backing out the quartz specific stub winsys since we can simply
use the generic stub winsys on quartz until we develop a standalone
winsys. Since we plan on removing all special cases for the stub winsys
by handling with a winsys vtable like all the others it's better if we
don't introduce a quartz specific stub.