Revert "Initial build of a Quartz winsys stub"
This reverts commit b2e41f1bfa53fe316bd42228fa26fc1d6dab5e9e.

We are backing out the quartz specific stub winsys since we can simply
use the generic stub winsys on quartz until we develop a standalone
winsys. Since we plan on removing all special cases for the stub winsys
by handling with a winsys vtable like all the others it's better if we
don't introduce a quartz specific stub.